requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
openai
faiss-cpu
chromadb-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc